test_settings/settings.py
import environ

env = environ.Env()

DEBUG = True
USE_TZ = True
DATABASES = {
    'default': env.db(),
}
INSTALLED_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'custom_user',
]
MIDDLEWARE_CLASSES = (
    'django.contrib.sessions.middleware.SessionMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
)
TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [],
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [],
        },
    },
]
SECRET_KEY = 'not_random'
ROOT_URLCONF = 'test_settings.urls'
AUTH_USER_MODEL = 'custom_user.EmailUser'
