requirements-test.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements-test.in, then run this to update:
#
#    pip-compile requirements-test.in
#
coverage==4.0
django-environ==0.4.0
mccabe==0.3.1             # via pylama
mock==1.3.0
pbr==1.8.1                # via mock
pep257==0.6.0             # via pylama
pep8==1.6.2               # via pylama
pyflakes==1.0.0           # via pylama
pylama==7.0.6
six==1.10.0               # via django-environ, mock
